Show better git branch status in zsh prompt
Changed --tags to --all which indicates when a remote branch is checked
out. Changed ${hook_com[branch]} to $(git rev-parse --short HEAD) which
prevents ambiguous commit hashes from being used. The zsh string
replacement %%.* was removed because it truncated branch names with
dots. The original purpose of that was to remove the ellipsis at the end
of plain commit hashes in hook_com[branch]

That line was replaced with git describe --all --always, which provides
extra information about where a commit comes from if it's not a branch
or a tag. The old line is still commented in case the new line becomes
annoying for some reason...