Support ssh Host patterns for terminal color
Thanks tpope-host for the regex

---

Newly discovered Include parsing caveat:

If an Include line is in a Host block, ssh will only apply the included
file if the parent Host matches. Hosts defined in the Include file are
incorrectly parsed as top-level Hosts by ssh_host_color.

This mess is already 9 golfed lines of overengineering so I don't want
to fix it.